requirements.txt
fastapi
uvicorn[standard]
pytesseract
python-multipart
Pillow